Server_Side/requirements.txt
# FastAPI web framework for building REST API
# Version 0.104.1 provides async support and automatic API documentation
fastapi==0.104.1
# Uvicorn ASGI server to run FastAPI application
# [standard] includes additional dependencies for better performance
# Version 0.24.0 supports async operations and WebSocket
uvicorn[standard]==0.24.0
# Motor library for asynchronous MongoDB operations
# Version 3.3.2 provides async MongoDB driver built on top of pymongo
motor==3.3.2
# Pydantic library for data validation and serialization
# Using newer version with pre-built wheels for Python 3.14
# Version >=2.9.0 provides improved performance and type validation
pydantic>=2.9.0
# OpenAI Python library for AI integration (compatible with Groq API)
# Using version 2.x for Python 3.14 compatibility
# Version >=2.15.0 supports async operations and Groq API compatibility
openai>=2.15.0
# HTTP client library - compatible with openai 2.x and Python 3.14
# Version >=0.25.0 provides async HTTP client for API requests
httpx>=0.25.0
# Python-dotenv for loading environment variables from .env file
# Version 1.0.0 loads environment variables from .env file into os.environ
python-dotenv==1.0.0
# BSON library for MongoDB ObjectId handling (included with pymongo)
# Version 4.6.0 provides MongoDB driver with ObjectId support
pymongo==4.6.0
